requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Task Queue
celery==5.3.4
redis==5.0.1

# GitHub Integration
PyGithub==2.1.1
cryptography==41.0.7
PyJWT==2.8.0

# AI/LLM
langchain==0.0.340
openai==1.3.5
anthropic==0.8.0
transformers==4.35.2
crewai==0.1.0

# Code Analysis
pylint==3.0.2
black==23.11.0
isort==5.12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
